% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPP_at_IEW218.R
\docType{package}
\name{OPP.at.ANU}
\alias{OPP.at.ANU}
\alias{OPP.at.ANU-package}
\title{Supplementary materials for AH Sparks' presentation at ANU Open Science Day}
\description{
Supplementary materials for AH Sparks' presentation at ANU Open Science Day
}
\author{
Adam H Sparks and Emerson M Del Ponte
}
