% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPP_at_IEW218.R
\docType{package}
\name{OPP.at.IEW2018}
\alias{OPP.at.IEW2018}
\alias{OPP.at.IEW2018-package}
\title{Supplementary materials for EM Del Ponte and AH Sparks' presentation at the
International Epidemiology Workshop 12 in Norway, 2018.}
\description{
Supplementary materials for EM Del Ponte and AH Sparks' presentation at the
International Epidemiology Workshop 12 in Norway, 2018.
}
\author{
Adam H Sparks and Emerson M Del Ponte
}
